requirements.txt
Flask==2.1.0
pymongo[srv]
pymongo
dnspython
configparser
google-cloud-vision